Drop oro, use rhino
oro has been flagged as dead for quite some time by the Apache project itself,
and it was time to find a replacement. Here it is, and it's called rhino.

Even though the 1.6 Sun JDK does have an implementation of it, we elect to use
the real one: some people may not use Sun's JDK.