Desupport custom simple type values, part 2
We now have gotten rid completely of SchemaNodeFactory, since it was only there
to pass along the list of types (as names) and validator classes to a
SchemaNode.

But now the latter has all validators built in, and as custom simple types are
not supported at all, this also simplifies SchemaValidator a great deal.