Desupport custom simple type values, part 1
The draft says:

----
If the property is not defined or is not in this list, then any type of value is
acceptable.  Other type values MAY be used for custom purposes, but minimal
validators of the specification implementation can allow any instance value on
unknown type values.
----

It is highly likely that it will be used at all.

So, start with desupporting multiple validators per primitive types, allow only
one.