Report path and schema keyword on syntax validation
This relies on a new object ValidationContext, which is for now used to
instantiate syntax validators but which will also be used later on for keyword
validators.

Also, since the property name is automatically appended to the log message,
refactor m{in,ax}{Items,Length} validators to use a common subclass which takes
care of positive integers.