ExtendsValidator: do merge schemas
It _is_ needed. Considering:

----
{
    "items" [ { schema1here }, { schema2here } ],
    "extends": {
        "additionalItems": { otherschemahere }
    }
}
----

validation will fail if the extended schema isn't merged with the base one, as
it will require all instances to obey otherschemahere which is _not_ what is
required by the spec. So, implement schema merge.

The implementation is quite crude and validates n+1 schemas for
a schema with n extensions: the base schema and the merged schema. This is
because the implementation of merging is quite crude by itself in that it takes
the base schema and force feeds it all keys from the extended schema(s). Crude,
but it works.