Syntax: make SyntaxValidator an interface
For the immediate future where we want to be able to register validators, the
need for separating syntax validators from keyword validators require that the
Validator interface be specialized. Start with SyntaxValidator. This is just
only a marker interface, which some people consider bad practice, but I don't
see any other way to do that and it does serve the purpose anyway.

Also, some Javadoc updates.