Fix DisallowValidator's report on $ref loop -- not perfect
For this, two new methods have been introduced to ValidationReport: error() and
isError(). The first will clear all messages from the report and replace it with
the only error message (preceded with "FATAL: " right after the path), and the
second reports whether the status is set to ERROR. Right now, only RefValidator
uses that.

Also, when merging reports, the merging report will now unconditionally set its
status to the same as the merged report, which means it will be correctly set to
ERROR in the event the merged report is in such a state.

This, along with .isError() testing in DisallowValidator, now avoids wrong
reporting but is far from being perfect...