ValidationReport: change logic for .addMessage() and .mergeWith()
In the even where the current report is in error state, don't append messages.
Also, if it is not in error state but the other is, clear messages, put in the
merged report's message in instead (normally there should be only one), and set
the status to error.

I'm sure some holes remain in the logic...