Nearly complete coverage for ref failures
The only missing parts are testing null fragment failures (but how it works is
obivous) and testing that a second call to a URL doesn't try and grab a second
copy (I do wonder how I'll do this one... Maybe I'll leave that for v0.2).

Now, missing Javadoc and publish to 0.1. Promise.