SyntaxFactory: implement an "ignored" keyword feature
For this, introduce a new "unchecked" set which contains the set of keywords to
be completely ignored.

When unregistering a validator, the matching keyword is added to the unecked
set. When registering it, fail if a validator for this keyword already exists --
this means the user won't have unregistered the old one before.