Move individual factories out of ValidationContext
Put them in a new class named ValidatorFactory. It is this class, now, which is
in charge of providing validators. This will allow, in later steps, to avoid
providing syntax validators completely (or an always true one) when using
well-known schemas, and also to make it easier to implement different
JsonValidators for different versions of the spec.