ValidationContext: only record in ref lookups if called from a ref!
Consider such a simple test as:

{
    "type": "integer";
}

and:

    validator.validate("#", somedata);

The first validation succeeds, however the second will FAIL as .validate() will
record the path in ref lookups the first time, and when it does so, first checks
that the ref has not been looked up. But we don't want that when initiating
validation!

Work around this by adding a boolean parameter to .getValidator(): if set to
false, the pointer lookup is _not_ recorded. Fix tests accordingly, since lookup
detection now proceeds from the second lookup, not the first.