$ref validation: don't require it to be by itself (or paired with "required")
Nothing in the draft says it should be. It is up to the implementation to handle
that.

A discussion on the mailing list says that if $ref is present, then the
implementation should act as if it were alone. Which is not the case right now,
and this needs to be fixed.