FormatFactory: don't consider a non existing "format" specification as fatal
The current draft doesn't say what should happen in such a case, and the choice
so far was to fail validation in such a case.

This is not the best thing to do. The default is now to ignore validation
entirely, completed by a log message at the warning level.

Modify documentation and tests to reflect these changes.