Introduce JsonRefException
OK, this is quite ugly. This exception is unchecked... And it is only launched
from what is still the $ref keyword validator.

Anyway, this does the job for now and it only required temporarily getting rid
of one test, and modifying three others.

However, one test also mysteriouly fails only on full runs :/ Add FIXME