Introduce JsonResolverBuilder
A resolver will also need a registry.

For now, there is one failing test with a cross-referencing $ref since the
registry is not there yet.

Also, SchemaContainer needs a new constructor with a URI as a supplementary
argument (case when a schema is downloaded from a URI but the resulting schema
does not have an id).