0.5.0beta3, and work around maven-release-plugin's stupidity
Namely: this idiotic plugin clones the entire repository again to perform a
release (it doesn't know about git archive, it seems) and creates nonsense
commits in the process.

I don't want either of those to happen, so .git/config edit by hand it is to use
a local bare clone instead.