JsonRef, SchemaContainer: normalize URIs behind the scene
Prior to this, an exception was thrown if the URI of a SchemaContainer was not
normalized. And this was the sole purpose of existence of JsonRef's
.isNormalized() method.

That was a little too far fetched. We need normalized URIs for resolution
anyway, so just normalize behind the scenes and don't bother the user with it.