Change interpretation of "hostname" and "email" keywords (FUNCTIONAL CHANGE)
If RFCs are strictly observed, emails may have no domain parts and host names
may not be fully qualified. However, this contradicts the general perception,
and 99+% use cases in the process.

From now on, this implementation expects hostnames to be fully qualified, and
email addresses to have a domain part.

This fixes issue #12.

v2: also modify javadoc for HostnameFormatSpecifier

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>