Consider schemas with relative IDs as anonymous (FUNCTIONAL CHANGE)
Before this commit, a parent schema whose id was not absolute threw an
exception.

Now, the implementation chooses to return an anonymous schema in such cases.
Note that a schema id is considered absolute _both_ if the URI is absolute and
the fragment part is either empty or non existent.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>