KeywordValidator: add first infrastructure elements for better reporting
All keywords are now required to supply their own name to the KeywordValidator
constructor -- and ideally implement .toString(). Special case failures to build
a keyword validator at all: in this case, the "keyword name" is considered to be
the class name.

Right now, no keyword implements .toString() at all, and the default
implementation is explanative enough...

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>