URIManager: implement schema URI redirection
It may happen that you want to be able to follow references to an external
schema which you already have locally. In this case, you may want to tell the
API that you want this external reference to be internal instead.

Implement the internal API part: in URIManager, add a translation map of source
URIs to destination URIs, and if the user asks to fetch from a source URI, fetch
the content fro the destination URI instead.

Add a (heavily commented) test to ensure that this mechanism works.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>