Remove remainder information from divisibleBy report message
This piece of information has turned out to be a programming headache. And
ultimately, it is not that useful.

Remove it, remove assorted hacks. People shoud figure out from themselves why a
given numeric instance is not a multiple of a divisor, if not by themselves, at
least using any calculator that one can find on any decent OS anyway.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>