Issue #14: only push/pop the context's container where needed
And that is only in one place: InstanceValidator.

Prior to this, it was needlessly pushed by "the four problem keywords",
ObjectValidator and ArrayValidator, but this led to a bug, highlighted by issue

As InstanceValidator is the only entrypoint into validation, only push and pop
in there: this cures the bug.

[fge: backported to 0.5.x]

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>