Fix bug in array element schema calculation when "items" is missing
If "items" is missing, it is equivalent to an empty schema.

This means that with this schema:

{
    "additionalItems": { "type": "string" }
}

any element of an array instance would have to obey the empty schema ({}), NOT
the schema in "additionalItems".

Add tests. And use Guava's @VisibleForTesting annotation while we are at it
(again another useful tool -- Guava really rocks).

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>