1.0.2
The implementation was incorrectly considering that in the absence of the
"items" keyword, then the schema for an array element should obey the schema in
"additionalItems". This is not the case however: no "items" is equivalent to
this keyword having an empty schema, which should be picked instead.

Francis Galiegue (1):
      Fix bug in array element schema calculation when "items" is missing

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>