NodeAndPath: add a "computed" boolean instance member
The initial plan with regards to path generation was to create an illegal
reference token in the path string.

However, this leads to potential special casing etc within code which will
eventually render it difficult to comprehend. Try another approach instead, that
is, add to NodePath the ability to tell whether a particular instance is the
result of a calculation and not an actual path in the schema. This way, when
this piece of information makes it into validation messages, the user will kknow
right away whether the actual schema path exists.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>