KeywordFactory: don't add validators which always validate successfully
For this, add an .alwaysTrue() method to KeywordValidator, and once the
validator is built, call this method and check whether it returns true (it
returns false by default).

Implement this method for the following keywords: additionalItems,
additionalProperties, properties, type. Other keywords could see this
implemented as well (for instance, minItems with default value 0), but more
often than not, schemas having these keywords will use a value which is not
obviously true.

Some validation code for these keywords could also do with simplification as a
result.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>