JarNamespaceValidationTest: devise test which does not work, but should
The test which breaks is due to a difference in behavior between:

    factory = JsonSchemaFactory.defaultFactory();
    schema = factory.fromURI("jar:/foo.jar!/a/b.json");

and:

    factory = new JsonSchemaFactory.Builder()
        .setNamespace("jar:/foo.jar!/a/").build();
    schema = factory.fromURI("b.json");

While the first solution will use the base URIDownloader and succeed, the second
one will fail since SchemaRegistry does not use JsonRef, which has a hack for
jar relative resolution, but URI resolution -- and resolving any URI against an
opaque URI, which "jar" URIs are, leads to the URI given as an argument.

First highlighted in issue #19.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>