Introduce JsonLocator class and implementations, plus tests
This class aims to replace all URIs used in various places for now to
materialize the non fragment part of a JSON Schema location.

Three implementations exist: the classical one, using fully RFC compliant URI
resolution schemes, an empty one for the empty locator since resolving it is
simple (it is always the resolved URI which is returned), and a dedicated "jar"
locator.

It could be extended to other opaque URIs quite easily.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>