Remove .asMap() method from custom ObjectNode implementations
Issue #30 shows that this can lead to a problem I would never had foreseen until
I saw it with my own eyes: the fact that the new ObjectNode implementation adds
an .asMap() method can cause problems.

As a result, re-introduce a JacksonUtils class with a sole asMap() method which
does the same thing.

Jackson is to blame here: ContainerNode<T> returns subclasses of itself in only
two methods. Two. But that is enough so that I must override ObjectNode and
ArrayNode this way, with this kind of problems lurking behind :/

Link: https://github.com/fge/json-schema-validator/issues/30
Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>