Rename SyntaxValidationResults; make it implement ProcessingReport
This way we need not set the level by hand.

The problem, though, is that we cannot inherit the log level settings of the
processing report passed as an argument since the cache loading function cannot
know it (it is initialized at build time). Which is why the configured log
levels are at "maximum values" (debug for logging, fatal for error level, fatal
for exceptions).

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>