Add "enum" as the last common keyword
"format" will be handled in yet another processor -- which will be made
optional.

"enum" is quite a challenge there. consider the following as values for this
keyword:

- [ 1, 2 ]
- [ 2, 1 ]
- [ 1.00, 2.0 ]
- [ 20e-1, 0.1e2 ]

All of these are _equal enums_.

Those will be handled correctly at validation time thanks to
JsonSchemaEquivalence, so no problem here. But the problem is how to build a
common digest for all of these -- and in a defined order, since this is
ultimately an array, and order matters when computing equality.

Meh. For now, just forget it, and pass the enum node as it is, and let the
keyword validator handle the differences. Still, it is an interesting problem.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>