Change exception construction in AbstractProcessingReport
Instead of throwing the exception itself in doLog(), use an overridable method
named doException().

Provide a first use of it in SyntaxProcessor, which will now throw an
InvalidSchemaException in its subreports.

OK, the syntax is a little ungainly, but a better syntax for that will have to
wait for Java 8.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>