URIManager: make it use LoadingConfiguration
LoadingConfiguration has everything needed to build a URI manager, so use it
instead, and remove all mutability methods from it.

Also add an .asMap() method to Dictionary as a side effect. This allows to
reimplement .entries in a safer manner -- and ultimately to get rid of methods
only used in SyntaxChecker.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>