Draft v3 core schema: disallow custom simple types
Note that this BREAKS v3, since section 5.1 says:

> Other type values MAY be used for custom purposes, but minimal validators of
> the specification implementation [sic] can allow any instance value on unknown
> type values.

But since the implementation already broke this section by declaring unknown
simple types are illegal, this is no big deal.

The change is now that _instance validation_ of a schema with that new meta
schema will cause a failure where it did not before. Again, that is a deliberate
break of the specification, but a useful one in any case: who creates custom
types when there is $ref?

Solves issue #40.