Change SchemaWalker
Issue #41 continued.

While not as generic as it could be (it will probably get more generic than that
later), the need here is to process the current schema before going any further.

Make SchemaWalker abstract. Require implementations to implement one method,
processCurrent(ProcessingReport, SchemaTree), so that whatever is needed to be
"done to the schema" is done _before_ walking further.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>