Modify pom.xml for OSGi-capable deployments
This patch does two things:

* adds the Apache Felix Bundle plugin to the pom file which will generate the
  appropriate OSGi headers;
* change the packaging type from 'jar' to 'bundle' to enable the bundle plugin to
  join the build.

'bundle' is the same process as 'jar' with the extra step of scanning the
classes and generating the OSGi headers into META-INF/MANIFEST.MF.  The only
difference between 'bundle' and 'jar' are the extra headers in the manifest.

[mbishop: I marked a few packages as optional because they are used in specific
types of attributes that may not be used. If a system wants to use those types
(like phone number) they can add the appropriate jars to their OSGi environment
and be good to go.]

[fge: modify commit message]