New class InstanceValidator
The existing ValidationProcessor is not easily modifiable to fulfill the needs
highlighted by issue #102. A ThreadLocal may be used, but that would make the
code rather convoluted.

Instead, create a new InstanceValidator class with the same signature; the
ValidationProcessor (which is unique per JsonSchemaFactory) will then create an
instance of that class for each attempted schema/instance validation pair.

This means the InstanceValidator instance will be what keyword validators will
use, instead of the ValidationProcessor.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>