Add tests to document behavior of '.' in pattern validation
The library treats "." as a wildcard even though the json validation spec is slightly ambiguous on certain regex special characters such as the period.  The expanded usage of the period as a wildcard is not in violation of the spec, but it could cause interop issues in the future, so additional documentation on its current behavior is being added via additional tests.